fix: wallet button responsive
